Add tests to SolarSystem, PlanetTester, SolarSystemTester, and UniverseTester so there's more coverage and edge case testing.
